+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zoo-code/types
 
+## Unreleased
+
+### Patch Changes
+
+- Version and document the host-webview protocol contract used by upcoming portable-core SDK adapter paths.
+
 ## 0.0.1
 
 ### Patch Changes